+++ tests/unit/test_validation_builder.py
@@ -314,67 +314,3 @@ def test_validation_add_filter(module_under_test, filter_field: dict):
     builder = module_under_test.ValidationBuilder(mock_config_manager)
     # Not currently checking effects, just that each filter type is accepted.
     builder.add_filter(filter_field)
-
-
-@pytest.mark.parametrize(
-    "input_list,max_length,expected_result",
-    [
-        (
-            [],
-            3,
-            [],
-        ),
-        (
-            [1, 2],
-            3,
-            [
-                [1, 2],
-            ],
-        ),
-        (
-            [1, 2, 3],
-            3,
-            [
-                [1, 2, 3],
-            ],
-        ),
-        (
-            [1, 2, 3],
-            2,
-            [
-                [1, 2],
-                [3],
-            ],
-        ),
-        (
-            [1, 2, 3],
-            3,
-            [
-                [1, 2, 3],
-            ],
-        ),
-        (
-            [1, 2, 3, 4],
-            3,
-            [
-                [1, 2, 3],
-                [
-                    4,
-                ],
-            ],
-        ),
-        (
-            [1, 2, 3, 4, 5, 6],
-            3,
-            [
-                [1, 2, 3],
-                [4, 5, 6],
-            ],
-        ),
-    ],
-)
-def test_list_to_sublists(
-    module_under_test, input_list: list, max_length: int, expected_result: list
-):
-    result = module_under_test.list_to_sublists(input_list, max_length)
-    assert result == expected_result
